feat(ui): InvoicePanel with invoice subnav; job-scoped invoice routes + shim
Mirrors the estimate panel extraction (Task 7): InvoiceDetailPage's guts
move into InvoicePanel.svelte, wired to job-scoped routes
(/jobs/:jobId/invoice[/:docId]) via JobInvoicePage, with the invoice
subnav sourced from /api/invoices/?job= (no change-order entries) and a
can_manage-gated "Start Invoice" empty state. The old /invoices/:id route
becomes a 12-line redirect shim. $canManageFinancials gating and all
seed/send-gate/adjustment/billables behavior preserved exactly; those
assertions move from InvoiceDetailPage's old tests into InvoicePanel's.
